fix: fix logic on subviews require lifecycle
The root cause is in ArrayRepeatStrategy, moveView and updateOverrideContext are two steps, but reorderable-repeat needs to update dnd resources after overrideContext is updated. Also, with moveView, all dnd resources need to be re-registered because of index changes, that also affects styling (css classes). More than that, when using reordering-group, addView/removeView could be called too, so need to move re-registration logic to right after instanceChanged(), with a delay to cater overrideContext update.

closes #11